ATLAS-5293: ATLAS UI: Add repo-wide Git hooks and dashboard pre-commit/push verification
